+++ config.js
@@ -1,0 +1,11 @@
+const config = {
+  MONGO_URL:
+    "mongodb://rainmongo:OSSawJsHkv85Hv6GA9uAKbEKtEX0bMlnyJ8W8HOEcHHudLsf0CAsjTtOdxY9zZfhuHT3UTMExsDMnCsOpJQG4Q==@rainmongo.mongo.cosmos.azure.com:10255/?ssl=true&replicaSet=globaldb&retrywrites=false&maxIdleTimeMS=120000&appName=@rainmongo@",
+  JWT_SECRET: "dev",
+  PORT: 5100,
+  MAIL_SERVICE: "gmail",
+  SENDER_MAIL: "rpmongotest@gmail.com",
+  MAIL_PASSWORD: "qpomthjiesygydjn",
+};
+
+module.exports = config;
