+++ helpers/Hashing.js
