--- src/config.js
@@ -1,10 +1,0 @@
-const config = {
-  MONGO_URL: "mongodb://localhost:27017/Rain",
-  JWT_SECRET: "dev",
-  PORT: 5100,
-  MAIL_SERVICE: "gmail",
-  SENDER_MAIL: "rpmongotest@gmail.com",
-  MAIL_PASSWORD: "qpomthjiesygydjn",
-};
-
-module.exports = config;
